Implement q_sort function
I tried implementing a merge sort algorithm for the doubly linked list,
but found it to be too complex to manage the previous and next pointers
during each merge or split operation. So I decided to break the loop in
the linked list and apply the merge sort algorithm as if it were a
standard singly linked list. After sorting, I relinked the previous
pointer for each element. I'm not sure if this approach is good, but it
simplifies the process for me.