Implement q_swap function
This time, I managed the prev pointer and next pointer in every
operation. The l1 pointer represents the first node we need to
swap in an operation, and the l2 pointer represents the second.
There are 6 steps to perform in each operation:

Change the next pointer of the node before l1 to point to l2.
Change the prev pointer of the node after l2 to point to l1.
Change the prev pointer of l2 to point to the prev pointer of l1.
Change the prev pointer of l1 to point to l2.
Change the next pointer of l1 to point to the next pointer of l2.
Change the next pointer of l2 to point to l1

Noted, the order of operations is crucial, and it's essential
to ensure that each operation is performed in the correct sequence.