Implement q_ascend and q_descend function
The q_ascend function removes every element that has a smaller element
after it. I use the min variable to record the minimum element and a
for loop to traverse all nodes from back to head. If the current node
is bigger than the min node, it means it has a smaller node after it,
so the current node is deleted. Otherwise, the current node is the
smallest node, and the min variable is updated to the current node.
After the function completes, the list will be in ascending order.
The q_descend function uses the same method but maintains the
biggest element instead.