Fix the problem of q_remove_tail and q_remove_head
I discovered a problem with these two functions: when the size of the
removed node's value is larger than the buffer size sp buffsize, it
doesn't terminate the string with '\0'. To address this issue, I fixed
it by ensuring that a '\0' character is inserted at the end of the string.