Update q_delete_mid
This problem can also be solved using the fast-slow-pointers method. In
this approach, the fast pointer moves two steps in every iteration,
while the slow pointer moves one step. When the fast pointer reaches the
end of the list and wraps around to the head pointer, the slow pointer
will be at the middle of the list. Freeing the memory pointed to by the
slow pointer completes this task.